[Feat]내비게이션 수정, 빈 화면일 때 문구 보여줌 (#90)
* feat: 로그인 외부 브라우저에서 처리

* feat: 내비게이션 수정, 빈 화면일 때 문구 보여줌

+++ data/src/main/java/org/ktc2/cokaen/wouldyouin/data/model/CurationResponse.kt
@@ -25,7 +25,7 @@ data class CurationResponse(
 
 
 data class CurationCuratorResponse(
-    val curatorId: Int,
+    val curatorId: Long,
     val nickname: String,
     val email: String,
     val phone: String,
